use concurrent workers to download using rest API
